Update cf-update.sh
o-o.myaddr.l.google.com was reporting back the wrong IP address for me, which I believe is because the ISP I use has made the switch to IPv6. 

This API service returns only the IPv4 address and nothing else. It is fast and has been working for many years.

Could the code perhaps be updated to allow for IPv6? I have to hold my hands up and say that I have some learning to do before I could manage that.